// src/types/DummyDataItem.ts
export type DummyDataItem = {
    IM_INCIDENT_KEY: number;
    FIRE_BOX: number;
    INCIDENT_TYPE_DESC: string;
    PROPERTY_USE_DESC: string;
    STREET_HIGHWAY: string;
    ZIP_CODE: number;
    BOROUGH_DESC: string;
    FLOOR: string;
    INCIDENT_MONTH: number;
    INCIDENT_DAY_OF_WEEK: number;
    INCIDENT_HOUR: number;
    RESPONSE_TIME: number;
    RESPONSE_TIME_MINUTES: number;
    YEAR: number;
    LATITUDE: number;
    LONGITUDE: number;
};